+++ src/engines/event-contract-collection-runner-runtime/index.ts
@@ -3,5 +3,6 @@ export * from "./EventContractCollectionRunnerScheduler";
 export * from "./EventContractCollectionRunnerFixtureWorker";
 export * from "./EventContractCollectionRunnerRuntimeOperator";
 export * from "./EventContractCollectionRunnerRuntimeAssemblyPlanner";
+export * from "./EventContractCollectionRunnerRuntimeForegroundStep";
 export * from "./EventContractCollectionRunnerRuntimeOwnershipRecovery";
 export * from "./EventContractCollectionRunnerRuntimeOwnershipQuarantine";